remove t-alis comonent
