--- src/App.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-import CreateExercise from './components/CreateExercise';
-
-function App() {
-  return (
-    <div>
-        <CreateExercise />
-    </div>
-  );
-}
-
-export default App;
